Extract Package into gfaas lib private interface
